# .specs/contracts/openapi.yaml
openapi: 3.0.4
info:
  title: Rampa Remittances API
  description: |
    API for Solana-based remittances app targeting Spain/Germany to LATAM corridor.
    
    ## Authentication Flow
    
    This API uses Web3Auth for user authentication with a token exchange pattern:
    
    1. **User Login**: User logs in via Web3Auth (Google, Apple, etc.) on the frontend
    2. **Token Exchange**: Frontend calls `/auth/web3auth/validate` with Web3Auth JWT token
    3. **API Token**: Backend validates Web3Auth token and returns our API JWT token
    4. **API Calls**: Frontend uses our API JWT token for all subsequent business API calls
    
    ## Security
    
    - All business endpoints require our API JWT token (Bearer authentication)
    - Web3Auth endpoints (`/auth/web3auth/*`) are public and don't require authentication
    - Health and inquiry endpoints are public
  version: 1.0.0
  contact:
    name: Rampa Team
    email: support@rampa.com
servers:
  - url: https://api.rampa.com
    description: Production server
  - url: https://staging-api.rampa.com
    description: Staging server
  - url: http://localhost:3001
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /onramp/initiate:
    post:
      summary: Initiate fiat to crypto conversion
      tags: [OnRamp]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OnRampRequest'
      responses:
        '201':
          description: On-ramp initiated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnRampResponse'
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
        '403':
          description: Profile verification required for this operation
  /onramp:
    get:
      summary: Get all on-ramps
      tags: [OnRamp]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, processing, completed, failed]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: On-ramps retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnRampResponse'
        '401':
          description: Unauthorized
  /onramp/pending:
    get:
      summary: Get pending on-ramps
      tags: [OnRamp]
      responses:
        '200':
          description: Pending on-ramps retrieved
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/OnRampResponse'
        '401':
          description: Unauthorized
  /onramp/stats/summary:
    get:
      summary: Get on-ramp statistics
      tags: [OnRamp]
      parameters:
        - name: startDate
          in: query
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: On-ramp statistics retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalOnRamp:
                    type: number
                  totalFees:
                    type: number
                  completedOnRamp:
                    type: number
                  failedOnRamp:
                    type: number
        '401':
          description: Unauthorized
  /onramp/{id}:
    get:
      summary: Get specific on-ramp
      tags: [OnRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: On-ramp retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnRampResponse'
        '404':
          description: On-ramp not found
        '401':
          description: Unauthorized

    post:
      summary: Process on-ramp
      tags: [OnRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerTransactionId]
              properties:
                providerTransactionId:
                  type: string
      responses:
        '200':
          description: On-ramp processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  providerTransactionId:
                    type: string
                  completedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /onramp/{id}/process:
    post:
      summary: Process on-ramp
      tags: [OnRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [providerTransactionId]
              properties:
                providerTransactionId:
                  type: string
      responses:
        '200':
          description: On-ramp processed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  providerTransactionId:
                    type: string
                  completedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /onramp/{id}/fail:
    post:
      summary: Fail on-ramp
      tags: [OnRamp]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [failureReason]
              properties:
                failureReason:
                  type: string
      responses:
        '200':
          description: On-ramp failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  status:
                    type: string
                  failureReason:
                    type: string
                  failedAt:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '401':
          description: Unauthorized
  /onramp/providers/{provider}/transaction/{providerTransactionId}:
    get:
      summary: Get on-ramp by provider
      tags: [OnRamp]
      parameters:
        - name: provider
          in: path
          required: true
          schema:
            type: string
        - name: providerTransactionId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: On-ramp retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OnRampResponse'
        '404':
          description: On-ramp not found
        '401':
          description: Unauthorized
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Our API JWT token obtained from /auth/web3auth/validate
  schemas:
    TokenType:
      type: string
      enum: [USDC, EURC, SOL]
...
    OnRampRequest:
      type: object
      required: [amount, fiatCurrency, tokenType]
      properties:
        amount:
          type: number
          format: decimal
          minimum: 10
        fiatCurrency:
          type: string
          enum: [EUR, USD]
        tokenType:
          $ref: '#/components/schemas/TokenType'
        paymentMethod:
          type: string
          enum: [credit_card, sepa_transfer]
    OnRampResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        walletId:
          type: string
          format: uuid
        type:
          type: string
          enum: [onramp, offramp]
        amount:
          type: number
          format: decimal
        fiatAmount:
          type: number
          format: decimal
        fiatCurrency:
          type: string
        tokenType:
          $ref: '#/components/schemas/TokenType'
        status:
          type: string
          enum: [pending, processing, completed, failed]
        provider:
          type: string
        providerTransactionId:
          type: string
        exchangeRate:
          type: number
          format: decimal
        fee:
          type: number
          format: decimal
        createdAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
        failedAt:
          type: string
          format: date-time
        failureReason:
          type: string
...